requirements.txt
alembic==1.0.0
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.8.24
chardet==3.0.4
click==6.7
docker-py==1.10.6
docker-pycreds==0.3.0
Flask==1.0.2
Flask-Cors==3.0.6
Flask-Migrate==2.2.1
Flask-Script==2.0.6
Flask-Session==0.3.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.5
itsdangerous==0.24
Jinja2==2.10
lxml==4.2.5
Mako==1.0.7
MarkupSafe==1.0
python-dateutil==2.7.3
python-editor==1.0.3
requests==2.18.3
six==1.11.0
SQLAlchemy==1.2.11
urllib3==1.22
websocket-client==0.51.0
Werkzeug==0.14.1
WTForms==2.2.1
